+++ release_notes.md
@@ -240,7 +240,7 @@ use `CodecPipelinesAdapter` manually.
 5. [TH2-3546][TH2-3583] `INTERACTIVE_MODE` - global parameter was introduced.
 6. `Data.use_cache()` <- True by default.
 7. Added data methods to get cache files paths `Data.get_cache_filepath()` and `Data.get_pending_cache_filepath()`.
-8. [TH2-3467] Added set of functions for datetime manipulation.
+8. [TH2-3467] Added utils classes to convert timestamps.
 
 ## Improvements
 1. [TH2-3003] Added automatic attachment of example.py code in readme.md.